search vehicle for plate finished
